removed data loss warn

+++ Core/include/SdlEngine/draw.h
@@ -14,7 +14,7 @@ public:
     static void Point(SDL_Renderer* renderer, const Vector2& point, SDL_Color color);
     static void Line(SDL_Renderer* renderer, const Vector2& start, const Vector2& end, SDL_Color color);
     static void Rectangle(SDL_Renderer* renderer, const Vector2& position, float width, float height, SDL_Color color, bool fill = true);
-    static void TextureRect(SDL_Renderer* renderer, const std::string& filePath, const Vector2& position, float width, float height);
+    static void TextureRect(SDL_Renderer* renderer, const std::string& filePath, const Vector2& position, int width, int height);
 
     // Collider-related functions
     // Add collider-related functions here
